requirements.txt

opencv-python>=4.9.0
numpy>=1.24.0
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.8
albumentations>=1.4.0
hydra-core>=1.3.2
pyyaml>=6.0.1
scikit-image>=0.24.0
matplotlib>=3.8.0
onnx>=1.16.0
onnxsim>=0.4.36

streamlit>=1.37.0
Pillow>=10.3.0
